delete useless code from dropper_torpedo in prep for rewrite
- batteryStatus.hpp
- pressureStatus.hpp
this board controls the front light and the dopper servos it doesn't
need to know about battery or pressure status.

- publisher.hpp
nothing useful was being published so that was deleted. when I rewrite I
will see if there is anything that should be published.

- serialCommand.hpp
this is not the appropriate place for a serial interface so that was
deleted. it can be rewritten elsewhere if needed (I don't think it is)